% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.R
\name{element}
\alias{element_futebol_badge}
\title{Theme Elements for Image Grobs}
\usage{
element_futebol_badge(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  size = 0.5
)
}
\arguments{
\item{alpha}{The alpha channel, i.e. transparency level, as a numerical value
between 0 and 1. \code{1L} skips alpha channel modification for more speed.}

\item{colour, color}{The image will be colorized with this color. Defaults to
\code{NA_character_} which means no change of color at all. Use the
special character \code{"b/w"} to set it to black and white. For more information
on valid color names in ggplot2 see
\url{https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill}.}

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{size}{The output grob size as a \link[grid:unit]{grid::unit}. If given a numeric,
\code{cm} will be applied as unit.}
}
\value{
An S3 object of class \code{element}.
}
\description{
In conjunction with the \link[ggplot2:theme]{ggplot2::theme} system, the following \code{element_}
functions enable images in non-data components of the plot, e.g. axis text.
\itemize{
\item \code{element_futebol_badge()}: draws Brazilian Soccer team badges instead of their abbreviations.
}
}
\details{
The elements translate Brazilian Soccer team abbreviations into badge images.
}
\examples{
\donttest{
library(futebolplotR)
library(ggplot2)

team_abbr <- valid_team_names()

df <- data.frame(
  random_value = runif(length(team_abbr), 0, 1),
  team = team_abbr
)

# use badges for x-axis
ggplot(df, aes(x = team, y = random_value)) +
  geom_col(aes(color = team, fill = team), width = 0.5) +
  scale_color_futebol(type = "secondary") +
  scale_fill_futebol(alpha = 0.4) +
  theme_minimal() +
  theme(axis.text.x = element_futebol_badge())

# use badges for y-axis
ggplot(df, aes(y = team, x = random_value)) +
  geom_col(aes(color = team, fill = team), width = 0.5) +
  scale_color_futebol(type = "secondary") +
  scale_fill_futebol(alpha = 0.4) +
  theme_minimal() +
  theme(axis.text.y = element_futebol_badge())

}
}
\seealso{
\code{\link[=geom_futebol_badges]{geom_futebol_badges()}}, and \code{\link[ggpath:element_path]{ggpath::element_path()}}
for more information on valid team abbreviations, and other parameters.
}
