% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nba_logos.R, R/nbaplotR-package.R
\docType{data}
\name{GeomNBAlogo}
\alias{GeomNBAlogo}
\alias{futebolplotR}
\alias{futebolplotR-package}
\title{futebolplotR: Brazilian Soccer Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomNBAlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomNBAlogo
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize Brazilian Soccer analysis in 'ggplot2'.
}
\author{
\strong{Maintainer}: Mioto Bruno \email{brunomioto97@gmail.com}

}
\keyword{datasets}
\keyword{internal}
