% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futebolplotR-package.R, R/geom_futebol_logos.R
\docType{package}
\name{futebolplotR-package}
\alias{futebolplotR}
\alias{futebolplotR-package}
\alias{GeomFUTlogo}
\title{futebolplotR: Brazilian Soccer Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomFUTlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFUTlogo
}
\description{
A set of functions to visualize Brazilian Soccer analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/futebolplotR}
  \item Report bugs at \url{https://github.com/brunomioto/futebolplotR/issues}
}

}
\author{
\strong{Maintainer}: Mioto Bruno \email{brunomioto97@gmail.com}

}
\keyword{datasets}
\keyword{internal}
