% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futebolplotR-package.R, R/geom_futebol_badges.R
\docType{package}
\name{futebolplotR-package}
\alias{futebolplotR}
\alias{futebolplotR-package}
\alias{GeomFUTbadge}
\title{futebolplotR: Brazilian Soccer Badges Plots in 'ggplot2'}
\format{
An object of class \code{GeomFUTbadge} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFUTbadge
}
\description{
A set of functions to visualize Brazilian Soccer analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/futebolplotR}
  \item \url{https://brunomioto.github.io/futebolplotR/}
  \item Report bugs at \url{https://github.com/brunomioto/futebolplotR/issues}
}

}
\author{
\strong{Maintainer}: Bruno Mioto \email{brunomioto97@gmail.com}

Authors:
\itemize{
  \item Gabriela Junqueira \email{mg.junqueira@hotmail.com}
}

}
\keyword{datasets}
\keyword{internal}
