% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid Brazilian Soccer Team Abbreviations}
\usage{
valid_team_names()
}
\value{
A vector of type \code{"character"}.
}
\description{
Output Valid Brazilian Soccer Team Abbreviations
}
\examples{
# List valid Brazilian soccer team abbreviations
valid_team_names()
}
